DSEGOG-67 refactoring plot settings into separate components
Much easier to maintain and figure out what's going on
